new-cv/backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication and security
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Validation and serialization
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# Environment and configuration
python-dotenv==1.0.0

# File handling
python-docx==1.1.0
PyPDF2==3.0.1
pdfplumber==0.10.3
openpyxl==3.1.2

# HTTP client and web scraping
requests==2.31.0
httpx==0.25.2
beautifulsoup4==4.12.2
lxml==4.9.3
html5lib==1.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Logging and monitoring
structlog==23.2.0

# Rate limiting
slowapi==0.1.9
redis==5.0.1

# Background tasks
celery==5.3.4

# Date and time utilities
python-dateutil==2.8.2
